Create filelistings.txt
The program is intended to read specifically formatted data from a text file
 one input at a time. Three consecutive inputs will be referred to as a file.


This first integer indicates the number of entries in the directory (or 0 if the entry is a plain file). The string is the name of the directory/file, and the last integer is the size of the file. Directories will have a 0 for their size.